FoundationNodeConfigPanel
